+++ examples/scripts/spm_CMAES.py
@@ -1,7 +1,7 @@
 import pybop
 import numpy as np
-import matplotlib.pyplot as plt
 
+# Define model
 parameter_set = pybop.ParameterSet("pybamm", "Chen2020")
 model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
 
@@ -19,6 +19,7 @@ parameters = [
     ),
 ]
 
+# Generate data
 sigma = 0.001
 t_eval = np.arange(0, 900, 2)
 values = model.predict(t_eval=t_eval)
@@ -26,6 +27,7 @@ CorruptValues = values["Terminal voltage [V]"].data + np.random.normal(
     0, sigma, len(t_eval)
 )
 
+# Form dataset for optimisation
 dataset = [
     pybop.Dataset("Time [s]", t_eval),
     pybop.Dataset("Current function [A]", values["Current [A]"].data),
@@ -33,23 +35,27 @@ dataset = [
 ]
 
 # Generate problem, cost function, and optimisation class
-problem = pybop.Problem(model, parameters, dataset)
+problem = pybop.FittingProblem(model, parameters, dataset)
 cost = pybop.SumSquaredError(problem)
 optim = pybop.Optimisation(cost, optimiser=pybop.CMAES)
 optim.set_max_iterations(100)
 
+# Run the optimisation
 x, final_cost = optim.run()
 print("Estimated parameters:", x)
 
-# Show the generated data
-simulated_values = problem.evaluate(x)
+# Plot the timeseries output
+pybop.quick_plot(x, cost, title="Optimised Comparison")
 
-plt.figure(dpi=100)
-plt.xlabel("Time", fontsize=12)
-plt.ylabel("Values", fontsize=12)
-plt.plot(t_eval, CorruptValues, label="Measured")
-plt.fill_between(t_eval, simulated_values - sigma, simulated_values + sigma, alpha=0.2)
-plt.plot(t_eval, simulated_values, label="Simulated")
-plt.legend(bbox_to_anchor=(0.6, 1), loc="upper left", fontsize=12)
-plt.tick_params(axis="both", labelsize=12)
-plt.show()
+# Plot convergence
+pybop.plot_convergence(optim)
+
+# Plot the parameter traces
+pybop.plot_parameters(optim)
+
+# Plot the cost landscape
+pybop.plot_cost2d(cost, steps=15)
+
+# Plot the cost landscape with optimisation path and updated bounds
+bounds = np.array([[0.6, 0.9], [0.5, 0.8]])
+pybop.plot_cost2d(cost, optim=optim, bounds=bounds, steps=15)