+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # [Unreleased](https://github.com/pybop-team/PyBOP)
 
+## Features
+- [#114](https://github.com/pybop-team/PyBOP/issues/114) - Adds standard plotting class `pybop.StandardPlot()` via plotly backend
+- [#114](https://github.com/pybop-team/PyBOP/issues/114) - Adds `quick_plot()`, `plot_convergence()`, and `plot_cost2d()` methods
 - [#116](https://github.com/pybop-team/PyBOP/issues/116) - Adds PSO, SNES, XNES, ADAM, and IPropMin optimisers to PintsOptimisers() class
+- [#38](https://github.com/pybop-team/PyBOP/issues/38) - Restructures the Problem classes ahead of adding a design optimisation example
+
+## Bug Fixes
 
 # [v23.11](https://github.com/pybop-team/PyBOP/releases/tag/v23.11)
 - Initial release
 - Adds Pints, NLOpt, and SciPy optimisers
 - Adds SumofSquareError and RootMeanSquareError cost functions
-- Adds Parameter and dataset classes
+- Adds Parameter and Dataset classes